+++ includes/term3d.h
@@ -1,12 +1,12 @@
 #ifndef TERM3D_H
 # define TERM3D_H
-# define SCREEN_HEIGHT		48
-# define SCREEN_WIDTH		63
+# define SCREEN_HEIGHT		50
+# define SCREEN_WIDTH		65
 # define OBJ_HEIGHT			45
 # define OBJ_WIDTH			60
 # define PHI				0.02
 # include <stddef.h>
-typedef struct s_vect	t_vect;
+# include "vector.h"
 typedef struct
 {
 	size_t	size;
