+++ includes/utils.h
@@ -1,0 +1,8 @@
+#ifndef UTILS_H
+# define UTILS_H
+# include "term3d.h"
+
+void	*or_exit(void *allocated);
+void	print_points(t_points points);
+
+#endif
