+++ includes/wrapper.h
@@ -1,5 +1,6 @@
 #ifndef WRAPPER_H
 # define WRAPPER_H
+# include <stdio.h>
 enum
 {
 	SYS_OK = 0,
