+++ libft/srcs/utils/ft_max_dbl.c
@@ -1,0 +1,8 @@
+
+double	ft_max_dbl(double a, double b)
+{
+	if (a > b)
+		return (a);
+	else
+		return (b);
+}
