+++ libft/srcs/utils/ft_min_dbl.c
@@ -1,0 +1,8 @@
+
+double	ft_min_dbl(double a, double b)
+{
+	if (a > b)
+		return (b);
+	else
+		return (a);
+}
