+++ srcs/parse/scaling.c
@@ -1,53 +1,21 @@
 #include <stdbool.h>
 #include <math.h>
-#include "input.h"
+#include "ft_utils.h"
+#include "parse.h"
 #include "term3d.h"
 #include "vector.h"
 
-double	ft_max_dbl(double a, double b)
-{
-	if (a > b)
-		return (a);
-	else
-		return (b);
-}
-
-double	ft_min_dbl(double a, double b)
-{
-	if (a > b)
-		return (b);
-	else
-		return (a);
-}
-
-void	get_max_or_min(t_points *points, t_vect *dest, t_cmp cmp)
-{
-	size_t	i;
-	t_vect	*vect;
-
-	*dest = points->points[0];
-	i = 0;
-	while (i < points->size)
-	{
-		vect = &points->points[i];
-		dest->x = cmp(dest->x, vect->x);
-		dest->y = cmp(dest->y, vect->y);
-		dest->z = cmp(dest->z, vect->z);
-		i++;
-	}
-}
-
-void	get_max_value(t_points *points, t_vect *max)
+static void	get_max_value(t_points *points, t_vect *max)
 {
 	get_max_or_min(points, max, ft_max_dbl);
 }
 
-void	get_min_value(t_points *points, t_vect *min)
+static void	get_min_value(t_points *points, t_vect *min)
 {
 	get_max_or_min(points, min, ft_min_dbl);
 }
 
-double	get_max_range(t_vect *max, t_vect *min)
+static double	get_max_range(t_vect *max, t_vect *min)
 {
 	t_vect	range;
 	double	max_range;
@@ -61,7 +29,7 @@ double	get_max_range(t_vect *max, t_vect *min)
 	return (max_range);
 }
 
-void	scale_points(t_points *points, double range)
+static void	scale_points(t_points *points, double range)
 {
 	const double	height_scale = OBJ_HEIGHT / range;
 	const double	width_scale = OBJ_WIDTH / range;
@@ -79,7 +47,7 @@ void	scale_points(t_points *points, double range)
 	}
 }
 
-void	scaling(t_points *points)
+void	scale_object_to_adjust_screen(t_points *points)
 {
 	double	range;
 	t_vect	max;
